fjernet unødvendig template story
