+++ packages/nav-frontend-chevron/stories/chevron.stories.tsx
@@ -1,21 +1,43 @@
 import React from "react";
-import NavFrontendChevron from "../src/chevron";
+import NavFrontendChevron, {
+  HoyreChevron,
+  VenstreChevron,
+  OppChevron,
+  NedChevron,
+} from "../src/chevron";
 import { Meta } from "@storybook/react/types-6-0";
 
 export default {
   title: "Chevron",
   component: NavFrontendChevron,
-  parameters: {
-    chromatic: { disable: true },
-  },
 } as Meta;
 
-const Template = ({ stor, type, ...args }) => (
-  <NavFrontendChevron type={type} stor={stor} {...args} />
-);
+export const All = () => (
+  <div
+    style={{
+      display: "grid",
+      gridAutoRows: "auto",
+      rowGap: "2rem",
+      gridAutoColumns: "fit-content",
+    }}
+  >
+    <div>
+      <HoyreChevron />
 
-export const Example = Template.bind({});
-Example.args = {
-  type: "høyre",
-  stor: false,
-};
+      <VenstreChevron />
+
+      <OppChevron />
+
+      <NedChevron />
+    </div>
+    <div>
+      <HoyreChevron stor />
+
+      <VenstreChevron stor />
+
+      <OppChevron stor />
+
+      <NedChevron stor />
+    </div>
+  </div>
+);
