+++ packages/nav-frontend-ekspanderbartpanel/stories/ekspanderbartpanel.stories.tsx
@@ -1,27 +1,90 @@
 import React from "react";
 import Ekspanderbartpanel from "../src/index";
-
+import { Normaltekst, Undertittel } from "nav-frontend-typografi";
+import Etikett from "nav-frontend-etiketter";
 import { Meta } from "@storybook/react/types-6-0";
 
 export default {
   title: "Ekspanderbartpanel",
   component: Ekspanderbartpanel,
-  parameters: {
-    chromatic: { disabled: true },
-  },
 } as Meta;
 
-const Template = ({ children, tittel, apen, border, ...args }) => (
-  <Ekspanderbartpanel border={border} apen={apen} tittel={tittel} {...args}>
-    {children}
-  </Ekspanderbartpanel>
-);
+export const All = () => (
+  <div
+    style={{
+      display: "grid",
+      gridAutoRows: "auto",
+      rowGap: "2rem",
+      gridAutoColumns: "fit-content",
+    }}
+  >
+    <Ekspanderbartpanel tittel="Default, klikk her for å åpne/lukke panelet">
+      Panelet vil da ekspandere og vise innholdet.
+    </Ekspanderbartpanel>
 
-export const Example = Template.bind({});
-Example.args = {
-  tittel: "Tittel",
-  apen: false,
-  border: true,
-  children:
-    "Adipisicing esse ut non id id. Aliqua reprehenderit in nostrud excepteur eiusmod dolor ullamco duis. Sint in tempor voluptate amet nulla qui cupidatat qui do anim est. Amet dolore Lorem aliqua nisi velit mollit. Fugiat qui nisi reprehenderit nostrud non. Fugiat duis occaecat proident exercitation est. Cupidatat dolor veniam duis aliqua ea adipisicing eu ea anim consequat magna aute.",
-};
+    <Ekspanderbartpanel
+      tittel={
+        <Normaltekst>
+          Endret tekststørrelse, klikk her for å åpne/lukke panelet
+        </Normaltekst>
+      }
+    >
+      Panelet vil da ekspandere og vise innholdet.
+    </Ekspanderbartpanel>
+
+    <Ekspanderbartpanel
+      tittel={
+        <div>
+          <Undertittel>
+            Avansert innhold i tittel, klikk her for å åpne/lukke panelet
+          </Undertittel>
+          <Etikett type="suksess">En etikett</Etikett>
+          <Etikett type="fokus" style={{ marginLeft: "0.5rem" }}>
+            En annen etikett
+          </Etikett>
+        </div>
+      }
+    >
+      Panelet vil da ekspandere og vise innholdet.
+    </Ekspanderbartpanel>
+
+    <Ekspanderbartpanel tittel="Nøstede ekspanderbartpaneler, klikk her for å åpne/lukke panelet">
+      Panelet vil da ekspandere og vise innholdet.
+      <Ekspanderbartpanel tittel="Klikk her for å åpne/lukke panelet">
+        Panelet vil da ekspandere og vise innholdet.
+      </Ekspanderbartpanel>
+    </Ekspanderbartpanel>
+
+    <Ekspanderbartpanel
+      apen
+      tittel="Default apen, klikk her for å åpne/lukke panelet"
+    >
+      Panelet vil da ekspandere og vise innholdet.
+    </Ekspanderbartpanel>
+
+    <Ekspanderbartpanel
+      apen
+      tittel="Nøstede ekspanderbartpaneler apen, klikk her for å åpne/lukke panelet"
+    >
+      Panelet vil da ekspandere og vise innholdet.
+      <Ekspanderbartpanel apen tittel="Klikk her for å åpne/lukke panelet">
+        Panelet vil da ekspandere og vise innholdet.
+      </Ekspanderbartpanel>
+    </Ekspanderbartpanel>
+
+    <Ekspanderbartpanel
+      border={false}
+      tittel="Uten border, klikk her for å åpne/lukke panelet"
+    >
+      Panelet vil da ekspandere og vise innholdet.
+    </Ekspanderbartpanel>
+
+    <Ekspanderbartpanel
+      border={false}
+      apen
+      tittel="Uten border apen, klikk her for å åpne/lukke panelet"
+    >
+      Panelet vil da ekspandere og vise innholdet.
+    </Ekspanderbartpanel>
+  </div>
+);
