+++ packages/nav-frontend-ikonknapper/stories/ikonknapper.stories.tsx
@@ -1,22 +1,111 @@
 import React from "react";
-import { Hamburgerknapp } from "../src/index";
+import {
+  Hamburgerknapp,
+  Menyknapp,
+  Nesteknapp,
+  Systemerknapp,
+  Søkeknapp,
+  Tilbakeknapp,
+  Xknapp,
+} from "../src/index";
 import { Meta } from "@storybook/react/types-6-0";
 
 export default {
   title: "Ikonknapper",
   component: Hamburgerknapp,
-  parameters: {
-    chromatic: { disable: true },
-  },
 } as Meta;
 
-const Template = ({ spinner, mini, kompakt, ...args }) => (
-  <Hamburgerknapp spinner={spinner} mini={mini} kompakt={kompakt} {...args} />
-);
+export const All = () => (
+  <div
+    style={{
+      display: "grid",
+      gridAutoRows: "auto",
+      rowGap: "2rem",
+      gridAutoColumns: "max-content",
+    }}
+  >
+    <div
+      style={{
+        display: "flex",
+        justifyContent: "space-evenly",
+      }}
+    >
+      <Hamburgerknapp />
 
-export const Example = Template.bind({});
-Example.args = {
-  spinner: false,
-  mini: false,
-  kompakt: false,
-};
+      <Menyknapp />
+
+      <Nesteknapp />
+
+      <Systemerknapp />
+
+      <Søkeknapp />
+
+      <Tilbakeknapp />
+
+      <Xknapp />
+    </div>
+
+    <div
+      style={{
+        display: "flex",
+        justifyContent: "space-evenly",
+      }}
+    >
+      <Hamburgerknapp mini />
+
+      <Menyknapp mini />
+
+      <Nesteknapp mini />
+
+      <Systemerknapp mini />
+
+      <Søkeknapp mini />
+
+      <Tilbakeknapp mini />
+
+      <Xknapp mini />
+    </div>
+
+    <div
+      style={{
+        display: "flex",
+        justifyContent: "space-evenly",
+      }}
+    >
+      <Hamburgerknapp kompakt />
+
+      <Menyknapp kompakt />
+
+      <Nesteknapp kompakt />
+
+      <Systemerknapp kompakt />
+
+      <Søkeknapp kompakt />
+
+      <Tilbakeknapp kompakt />
+
+      <Xknapp kompakt />
+    </div>
+
+    <div
+      style={{
+        display: "flex",
+        justifyContent: "space-evenly",
+      }}
+    >
+      <Hamburgerknapp mini kompakt />
+
+      <Menyknapp mini kompakt />
+
+      <Nesteknapp mini kompakt />
+
+      <Systemerknapp mini kompakt />
+
+      <Søkeknapp mini kompakt />
+
+      <Tilbakeknapp mini kompakt />
+
+      <Xknapp mini kompakt />
+    </div>
+  </div>
+);
